Create without-littering macro
Creates environment for using helper functions var and etc as well as a
namespace for variables related to each third-party package. The third
party package files are put into corresponding subdirectories if they
can be found using locate-library. Using locate-library we can decide
whether the package is installed using load-path without having to worry
about whether the package is loaded. This introduces the restriction
that the load path needs to be set up before no-littering is used.