Don't set desktop-base-{file,lock}-name
Previously we set the values of these variables to absolute file names,
but that was wrong because the values are supposed to be basenames only.
Also this should be hidden files because the user can save different
desktop files in different locations, e.g. at the top-level of some
project.