semantic: theme semanticdb-default-save-directory
If we followed the conventions strictly, then the directory should be
named "semanticdb-default-save", but that would be a silly name, so we
use "semantic" instead.  Using the default value of the option, the name
of the directory would be "semanticdb", but we don't use that because
the package is named "semantic" even though some symbols use the prefix
"semanticdb-".