epkg: theme epkg-repository
Do this for the benefit of users who use `epkg' but not `borg'.
For Emacs.g users this is not necessary because that already
sets this variable to "~/.emacs.d/var/epkgs/" in its `master'
branch.

Also I do realize that this doesn't quite follow the conventions.
That's because using `<prefix>/epkgs/' makes a lot of sense for
Emacs.g (which is not bound to these conventions) and it would be
very confusing if the location would change once an Emacs.g user
assimilated `no-littering'.