notmuch: theme notmuch-init-file
Notmuch's handling of its init file is needlessly complex, fragile
and confusing to users.  Also the observed behavior does not fully
correspond with the claimed behavior ('), and the justification for
the complexity is dubious (").

(') This is wrong: "If a file with one of the suffixes defined by
    `get-load-suffixes' exists, it will be read instead."
    It should read: "If and only if a file with one of the suffixes
    defined by `get-load-suffixes' exists, it will be read."

(") See the comment preceding `notmuch-init-file's definition.

The init file is loaded if and only if the variable that controls
its name and location claims that the file does NOT end with ".el"
and the file DOES end with ".el".  Otherwise `notmuch' silently
forgoes loading the file.

This is the code that determines whether the config file exists:

  (locate-file notmuch-init-file  '("/") (get-load-suffixes))

And these are some situation a user might arrive at:

1) Value of notmuch-init-file: "/path/to/notmuch-init"
   File "/path/to/notmuch-init.el" exists.
   File "/path/to/notmuch-init" does not exist.
   Above code returns: "/path/to/notmuch-init.el"

2) Value of notmuch-init-file: "/path/to/notmuch-init.el"
   File "/path/to/notmuch-init.el" exists.
   File "/path/to/notmuch-init" does not exist.
   Above code returns: nil

3) Value of notmuch-init-file: "/path/to/notmuch-init"
   File "/path/to/notmuch-init.el" does not exist.
   File "/path/to/notmuch-init" exists.
   Above code returns: nil

4) Value of notmuch-init-file: "/path/to/notmuch-init.el"
   File "/path/to/notmuch-init.el" does not exist.
   File "/path/to/notmuch-init" exists.
   Above code returns: nil

Situation (1) is the only one that works.