eshell: theme eshell-aliases file separately from other eshell files
The aliases file stores user-defined aliases used in Eshell. Users
would typically be inclined to check them into version control as they
would with snippets. The aliases are not defined using s-expressions.

The em-alias.el file normally resides inside of the directory
corresponding to eshell-directory-name, and does not assist in
generating an additional containing directory, so we must make that
directory ourselves.

https://git.savannah.gnu.org/cgit/emacs.git/tree/lisp/eshell/em-alias.el#n103