erc: theme erc-dcc-get-default-directory
For the ERC chat client, if erc-dcc-get-default-directory is nil (the
default), then the current value of default-directory is used for all
incoming DCC file transfers. We prefer to store all the files in
/var/erc/dcc instead.

ERC does not appear to make any effort to generate the specified
directory if it does not already exist, so we do it ourselves.

Defined here:
https://git.savannah.gnu.org/cgit/emacs.git/tree/lisp/erc/erc-dcc.el#n390