code-review: Theme variables
Theme `code-review-db-database-file', `code-review-download-dir', and
`code-review-log-file'.

https://github.com/wandersoncferreira/code-review/

code-review allows for code reviews to be performed within Emacs.

`code-review-db-database-file' stores the code-review database.

`code-review-download-dir' is a directory into which binary files are
downloaded. This directory is made by the code-review package if it
does not already exist.

`code-review-log-file' is a log file to which the code-review package
outputs certain warnings. The warnings are not in the form of
s-expressions.