harpoon: Theme harpoon-cache-file
Despite the name, this variable actually specifies the directory that
contains the cache-file.  The file-name of is hard-coded as "harpoon".
This results in "<var>/harpoon/harboon", which isn't great but
contains the issue inside that directory.

The use of (var "./") was proposed, but I don't want such an entry
appear in "no-littering.el" as that could lead to others thinking
that this is a reasonable thing to do, even in situations were there
are clearly better alternatives.

Closes #189.