detached: Theme detached-db-directory and detached-session-directory
`detached-db-directory` is used to store the detached database and log
files.  `detached-sessions-directory` is used to store log-files with
outputs of detached sessions.

Both directories are created automatically by detached if they don't
exist yet.

By default, the `detached-session-directory` is in a sub-path of
`detached-db-directory`, i.e. in my case the defaults were:

  detached-db-directory: ~/.emacs/detached/
  detached-session-directory: ~/.emacs/detached/session/

but I think that it seems more in line with the no-littering
conventions to split them into different sub-paths under var/detached.