readme: Update recentf hint
Since [1: c753e4cb2a2], recentf abbreviates file-names by default.
That is unfortunate because it is a backward incompatible change that
breaks code like this example.  It likely would have been possible to
only change how the file-names are displayed, without affecting the
format in which they are being compared.

Using `recentf-expand-file-name' should ensure that we provide file-
names in the format expected by recentf, regardless of what version of
recentf is in use, and thus what format is expects.

Closes #199.

1: 2022-06-23 c753e4cb2a2a5998626d48d85071c047cbab6c54
   Use shortened filenames in recentf-mode