Use variables in delayed calls to make-directory
Many packages allow theming the locations of their files, but fail to
take into account, that that means that directories may be used that
do not exist yet, so this package has to take care them, if necessary.

Previously we always duplicated the call to `etc' or `var', when doing
that; now we use the themed variable, if there is one that corresponds
directly to the directory.  That way, if a user uses `no-littering'
but additionally customized some themed variables, we don't end up
creating unused empty directories.  This was suggested in #204.

When there is no variable corresponding directly with the directory
that we are forced to create, then we might still end up creating an
unused empty directory.  To prevent that, upstream would have to take
care of creating its directories itself.

Additionally group the forms that create directories, separately from
the forms that set variables.  That makes it easier to see, which
packages fail to create their directories, and generally results in
more readable code/data.