shell-maker: Theme shell-maker-history-path
The path under which the history file directories for each `shell-maker`
implementation are stored. Notable shell-maker implementations are
`chatgpt-shell` and `dall-e-shell` by the same developer and contained in the
same git repository [1]. Their histories are
stored in

```
  <shell-maker-history-path>
  ├── chatgpt
  │   └── history
  └── dall-e-shell
      └── history
```

By default, it takes the value of `user-emacs-directory`. I didn't look into the
code but in my tests the directory was created if absent.

The history-files are simple text files containing the user prompt histories.

The `chatgpt-shell` package also creates the alias `chatgpt-shell-history-path`
to this variable.

[1]: https://github.com/xenodium/chatgpt-shell