tabspaces: Theme tabspaces-session-file
Tabspaces (https://github.com/mclear-tools/tabspaces) leverages the
tab-bar feature to create buffer-isolated workspaces called
tabspaces. The tabspaces-session-file is used for persisting tabspace
configuration across emacs restarts.  The default configuration is to
save the session file as 'tabsession.el' in the
user-emacs-directory. This is an s-expressions file.  This is the only
data file written by the package.