erc: Theme erc-startup-file-list
erc-startup-file-list is a list of .bashrc-like files that are
searched for by erc, which executes the first extant and readable one.

If the executed file ends with a ".el" extension, then it is loaded as
an Elisp script, while if it does not end with that extension, then it
is treated as a bunch of IRC commands.

Following the upstream convention, we put the Elisp variant prior to
the non-Elisp variant in the list.