server: Theme server-auth-dir
`server-auth-dir' is used to store text files containing authentication keys.
This is the only configuration file or directory of the package intended to be
user-modified. This package does take care of creating the containing directory
if necessary.

Clarification on `server-auth-dir' being the only path "intended to be
user-modified": there is a variable named `server-socket-dir' which names a
directory used to create sockets, but this package shouldn't change it. The
reason is that `server-socket-dir' is used by emacsclient to connect to the server
instance, and emacsclient doesn't read the init file, and so would never run
this package to learn about any new value. Additionally, the path is already out
of the way, being something like "/run/user/1000/emacs" on Linux.

We still put `server-auth-dir' in a subdirectory in case something changes about
that mechanism.