eshell: Create config directory
By default `eshell-aliases-file', `eshell-login-script',
`eshell-rc-script' and `eshell-history-file-name' are all placed
in `eshell-directory-name'.

The first three belong in "etc/" but the history file belongs in
"var/".  We have chosen to theme the first three individually and
theme the history file variable indirectly by theming the directory
variable.  (Probably assuming that any additional files that might
be placed in `eshell-directory-name' will also belong in "var/".

We used to take care to create `eshell-directory-name'.  That is
not necessary because `eshell-mode' takes care of that.  On the
other hand we need to ensure "etc/eshell/" exists, but failed to
do so.  That resulted in `eshell-write-aliases-list' silently
failing.

Like `eshell-mode' does for `eshell-directory-name', ensure that
"etc/eshell" uses safe permissions.

Closes #234.